requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bidict==0.23.1
bottle==0.13.4
certifi==2026.2.25
click==8.3.1
docutils==0.22.4
fastapi==0.135.3
fastexcel==0.19.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
lxml_html_clean==0.4.4
markdown2==2.5.5
MarkupSafe==3.0.3
multidict==6.7.1
nicegui==3.9.0
orjson==3.11.8
polars==1.39.3
propcache==0.4.1
proxy_tools==0.1.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-engineio==4.13.1
python-multipart==0.0.24
python-socketio==5.16.1
pywebview==6.1
PyYAML==6.0.3
simple-websocket==1.1.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
watchfiles==1.1.1
wsproto==1.3.2
yarl==1.23.0
